+++ lib/chargify/cash_report.rb
@@ -1,4 +1,5 @@
 require_relative "report"
+require_relative "charge"
 
 module Reports
   class CashReport < Report
@@ -155,178 +156,6 @@ module Reports
       else
         [customer.first_name, customer.last_name].join(" ")
       end
-    end
-
-    class ChargeRow
-      attr_reader :charge
-
-      def initialize(charge)
-        @charge = charge
-      end
-
-      delegate :company_name, :customer_reference, :statement_id,
-        :product_handle, :coupon_code, to: :charge
-
-      def row
-        [
-          created_at,
-          company_name,
-          customer_reference,
-          statement_id,
-          product_handle,
-          "payment",
-          coupon_code,
-          coupon_amount,
-          net_price,
-          discount,
-          net_after_discount,
-          tax_amount,
-          total
-        ]
-      end
-
-      def created_at
-        charge.created_at.to_s(:db)
-      end
-
-      def net_price
-        format_value(:net_price)
-      end
-
-      def discount
-        format_value(:discount)
-      end
-
-      def net_after_discount
-        format_value(:net_after_discount)
-      end
-
-      def tax_amount
-        format_value(:tax_amount)
-      end
-
-      def total
-        format_value(:total)
-      end
-
-      def coupon_amount
-        if charge.coupon_amount.nil?
-          ""
-        else
-          "%d%" % charge.coupon_amount
-        end
-      end
-
-      private
-
-      def format_value(value)
-        "%d" % (charge.send(value) / 100)
-      end
-    end
-
-    class Charge
-      extend Forwardable
-
-      attr_reader :transactions, :products, :customers
-
-      def initialize(transactions, products, customers)
-        @transactions = transactions
-        @products     = products
-        @customers    = customers
-      end
-
-      def to_row
-        ChargeRow.new(self).row
-      end
-
-      def obj
-        (transactions['Payment'] || transactions['Adjustment'] || transactions['Charge']).first
-      end
-
-      def adjustment?
-        obj.type == 'Adjustment'
-      end
-
-      def total
-        if adjustment?
-          charges_and_adjustments.sum(&:amount_in_cents)
-        else
-          obj.amount_in_cents
-        end
-      end
-
-      def net_price
-        charges['baseline'].first.amount_in_cents
-      end
-
-      def net_after_discount
-        net_price - discount.abs
-      end
-
-      def tax_amount
-        if charges['tax'].present?
-          charges['tax'].first.amount_in_cents.to_i
-        else
-          0
-        end
-      end
-
-      def charges
-        transactions['Charge'].group_by(&:kind)
-      end
-
-      def charges_and_adjustments
-        transactions.values.flatten.select do |transaction|
-          %w[Charge Adjustment].include?(transaction.type)
-        end
-      end
-
-      def discount
-        if adjustment?
-          obj.amount_in_cents
-        else
-          0
-        end
-      end
-
-      def_delegators :obj, :customer_id, :product_id, :statement_id, :created_at
-
-      def customer
-        @customers[customer_id]
-      end
-
-      def product
-        @products[product_id]
-      end
-
-      def company_name
-        if customer.organization.present?
-          customer.organization
-        else
-          [customer.first_name, customer.last_name].join(" ")
-        end
-      end
-
-      def customer_reference
-        customer.reference.to_s
-      end
-
-      def product_handle
-        product.handle
-      end
-
-      def coupon_code
-        return "" unless adjustment?
-
-        /Coupon: (.+) -/.match(obj.memo)[1]
-      end
-
-      def coupon_amount
-        return if discount.zero?
-
-        (discount.abs / (net_price / 100))
-      end
-
     end
   end
 end
